Add bootstrap.sh to launch the test in the CentOS CI
The `bootstrap.sh` script comes from gluster/centosci:scripts/common/
and is used to start a shell script as a test on the reserved machine
that is provided by the CentOS CI.

This script is used by the rewrite of the Jenkins Job from XML format to
a Jenkins Job Builder .yml file. It will be easier to update the
configuration this way, and provides for more re-use of standard
scripts.

Signed-off-by: Niels de Vos <ndevos@redhat.com>